refactor: extract recommendation fetcher hook
