+++ frontend/src/hooks/useRecommendations.ts
@@ -7,14 +7,13 @@ import {
   DEFAULT_ACTIVE_WEEK,
   DEFAULT_WEEK,
   RecommendationRow,
-  NormalizeRecommendationResult,
   buildRecommendationRows,
   formatWeekLabel,
-  normalizeRecommendationResponse,
 } from '../utils/recommendations'
 
-const week = weekModule as typeof import('../lib/week')
+import { useRecommendationFetcher } from './recommendationFetcher'
 
+const week = weekModule as typeof import('../lib/week')
 const api = apiModule as typeof import('../lib/api') & {
   fetchRecommend?: (input: { region: Region; week?: string }) => Promise<RecommendResponse>
 }
@@ -36,55 +35,6 @@ export interface UseRecommendationsResult {
   displayWeek: string
   sortedRows: RecommendationRow[]
   handleSubmit: (event: FormEvent<HTMLFormElement>) => void
-}
-
-interface RecommendationFetchInput {
-  region: Region
-  week: string
-  preferLegacy?: boolean
-}
-
-export type RecommendationFetcher = (
-  input: RecommendationFetchInput,
-) => Promise<NormalizeRecommendationResult | null>
-
-export const useRecommendationFetcher = (): RecommendationFetcher => {
-  return useCallback<RecommendationFetcher>(
-    async ({ region, week, preferLegacy = false }) => {
-      const callModern = async (): Promise<RecommendResponse | undefined> => {
-        if (typeof api.fetchRecommendations !== 'function') {
-          return undefined
-        }
-        try {
-          return await api.fetchRecommendations(region, week)
-        } catch {
-          return undefined
-        }
-      }
-
-      const callLegacy = async (): Promise<RecommendResponse | undefined> => {
-        if (typeof api.fetchRecommend !== 'function') {
-          return undefined
-        }
-        try {
-          return await api.fetchRecommend({ region, week })
-        } catch {
-          return undefined
-        }
-      }
-
-      const primary = preferLegacy ? callLegacy : callModern
-      const secondary = preferLegacy ? callModern : callLegacy
-
-      const response = (await primary()) ?? (await secondary())
-      if (!response) {
-        return null
-      }
-
-      return normalizeRecommendationResponse(response, week)
-    },
-    [],
-  )
 }
 
 const useCropIndex = (): Map<string, number> => {
@@ -297,3 +247,4 @@ export const useRecommendations = ({ favorites, initialRegion }: UseRecommendati
 }
 
 export type { RecommendationRow } from '../utils/recommendations'
+export type { RecommendationFetcher } from './recommendationFetcher'
